+++ cli/commands/root/root.go
@@ -1,6 +1,7 @@
 package root
 
 import (
+	colorable "github.com/mattn/go-colorable"
 	"github.com/sensu/sensu-go/cli"
 	"github.com/sensu/sensu-go/cli/client/config"
 	"github.com/sensu/sensu-go/cli/commands/version"
@@ -18,6 +19,10 @@ func Command() *cobra.Command {
 			_ = cmd.Help()
 		},
 	}
+
+	// Use colorable (fixes Windows color support)
+	cmd.SetOut(colorable.NewColorableStdout())
+	cmd.SetErr(colorable.NewColorableStderr())
 
 	// Templates
 	cmd.SetUsageTemplate(usageTemplate)